+++ Pillepalle1.StatefulApi/src/Customers.Application/Entities/Customer.cs
@@ -2,7 +2,7 @@
 
 public class Customer
 {
-    public required int Id { init; get; }
+    public required long Id { init; get; }
     public required string Name { init; get; }
     public required string Details { init; get; }
 }
